chore: moved navbar to website
