+++ apps/website/components/organisms/navbar/navbar-notifications.tsx
@@ -1,3 +1,5 @@
+import { useMenu } from '@gateway/ui';
+
 import NotificationsIcon from '@mui/icons-material/Notifications';
 import Avatar from '@mui/material/Avatar';
 import Badge from '@mui/material/Badge';
@@ -7,15 +9,11 @@ import MenuItem from '@mui/material/MenuItem';
 import Tooltip from '@mui/material/Tooltip';
 import Typography from '@mui/material/Typography';
 
-import { useMenu } from '../../hooks/use-menu';
-
-const settings = ['Profile', 'Account', 'Dashboard', 'Logout'];
-
 export function NavBarNotifications() {
   const userMenu = useMenu();
   return (
     <>
-      <Tooltip title="Open settings">
+      <Tooltip title="Open Notifications">
         <IconButton onClick={userMenu.onOpen}>
           <Badge color="secondary" variant="dot" overlap="circular">
             <Avatar sx={{}}>
@@ -40,11 +38,11 @@ export function NavBarNotifications() {
         open={userMenu.isOpen}
         onClose={userMenu.onClose}
       >
-        {settings.map((setting) => (
+        {/* {settings.map((setting) => (
           <MenuItem key={setting} onClick={userMenu.onClose}>
             <Typography textAlign="center">{setting}</Typography>
           </MenuItem>
-        ))}
+        ))} */}
       </Menu>
     </>
   );
