+++ libs/ui/src/index.ts
@@ -1,3 +1,2 @@
 export * from './lib/motion-components/motion-components';
-export * from './lib/navbar/navbar';
 export * from './hooks/use-menu';
